cleaned up my Fingo and ran go fmt on the entire coattails directory
